Init project with React-CDK
Storybooks runs on a separate pork and lets you view UI components interactively while they're being developed. Additionally, the makers of Storybooks, @Kadira, recommends publishing each component to NPM so they can be explicitly added as project dependencies. I'm still on the fence as to the value of doing this since our project scope is significantly smaller, but I will at least try this route since it's suggested and documented.